@@ -139,7 +139,7 @@ func TestDNSProvider_getHostedZone(t *testing.T) {
 
 	domains := []govultr.Domain{{Domain: "example.com"}, {Domain: "example.org"}, {Domain: "example.net"}}
 
-	for i := 0; i < 50; i++ {
+	for i := range 50 {
 		domains = append(domains, govultr.Domain{Domain: fmt.Sprintf("my%02d.example.com", i)})
 	}
 
@@ -157,7 +157,6 @@ func TestDNSProvider_getHostedZone(t *testing.T) {
 	}
 
 	for _, test := range testCases {
-		test := test
 		t.Run(test.desc, func(t *testing.T) {
 			t.Parallel()
 
