@@ -1,8 +1,6 @@
 package azuredns
 
 import (
-	"net/http"
-	"net/http/httptest"
 	"testing"
 	"time"
 
@@ -26,19 +24,9 @@ func TestNewDNSProvider(t *testing.T) {
 		expected string
 	}{
 		{
-			desc: "success",
-			envVars: map[string]string{
-				EnvEnvironment:    "",
-				EnvSubscriptionID: "A",
-				EnvResourceGroup:  "B",
-			},
-		},
-		{
 			desc: "unknown environment",
 			envVars: map[string]string{
-				EnvEnvironment:    "test",
-				EnvSubscriptionID: "A",
-				EnvResourceGroup:  "B",
+				EnvEnvironment: "test",
 			},
 			expected: "azuredns: unknown environment test",
 		},
@@ -63,78 +51,6 @@ func TestNewDNSProvider(t *testing.T) {
 			require.NotNil(t, p.provider)
 
 			assert.IsType(t, p.provider, new(DNSProviderPublic))
-		})
-	}
-}
-
-func TestNewDNSProviderConfig(t *testing.T) {
-	testCases := []struct {
-		desc           string
-		subscriptionID string
-		resourceGroup  string
-		privateZone    bool
-		handler        func(w http.ResponseWriter, r *http.Request)
-		expected       string
-	}{
-		{
-			desc:           "success (public)",
-			subscriptionID: "A",
-			resourceGroup:  "B",
-			privateZone:    false,
-		},
-		{
-			desc:           "success (private)",
-			subscriptionID: "A",
-			resourceGroup:  "B",
-			privateZone:    true,
-		},
-		{
-			desc:           "SubscriptionID missing",
-			subscriptionID: "",
-			resourceGroup:  "",
-			expected:       "azuredns: SubscriptionID is missing",
-		},
-		{
-			desc:           "ResourceGroup missing",
-			subscriptionID: "A",
-			resourceGroup:  "",
-			expected:       "azuredns: ResourceGroup is missing",
-		},
-	}
-
-	for _, test := range testCases {
-		t.Run(test.desc, func(t *testing.T) {
-			config := NewDefaultConfig()
-			config.SubscriptionID = test.subscriptionID
-			config.ResourceGroup = test.resourceGroup
-			config.PrivateZone = test.privateZone
-
-			mux := http.NewServeMux()
-			server := httptest.NewServer(mux)
-			t.Cleanup(server.Close)
-
-			if test.handler == nil {
-				mux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {})
-			} else {
-				mux.HandleFunc("/", test.handler)
-			}
-
-			p, err := NewDNSProviderConfig(config)
-
-			if test.expected != "" {
-				require.EqualError(t, err, test.expected)
-				return
-			}
-
-			require.NoError(t, err)
-			require.NotNil(t, p)
-			require.NotNil(t, p.provider)
-
-			if test.privateZone {
-				assert.IsType(t, p.provider, new(DNSProviderPrivate))
-			} else {
-				assert.IsType(t, p.provider, new(DNSProviderPublic))
-			}
 		})
 	}
 }
