@@ -9,40 +9,30 @@ import (
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/privatedns/armprivatedns"
 	"github.com/go-acme/lego/v4/challenge/dns01"
-	"github.com/go-acme/lego/v4/platform/config/env"
 )
 
 // DNSProviderPrivate implements the challenge.Provider interface for Azure Private Zone DNS.
 type DNSProviderPrivate struct {
-	config       *Config
-	zoneClient   *armprivatedns.PrivateZonesClient
-	recordClient *armprivatedns.RecordSetsClient
+	config                *Config
+	credentials           azcore.TokenCredential
+	serviceDiscoveryZones map[string]ServiceDiscoveryZone
 }
 
-// NewDNSProviderPrivate creates a DNSProviderPrivate structure with initialized Azure clients.
+// NewDNSProviderPrivate creates a DNSProviderPrivate structure.
 func NewDNSProviderPrivate(config *Config, credentials azcore.TokenCredential) (*DNSProviderPrivate, error) {
-	options := arm.ClientOptions{
-		ClientOptions: azcore.ClientOptions{
-			Cloud: config.Environment,
-		},
-	}
-
-	zoneClient, err := armprivatedns.NewPrivateZonesClient(config.SubscriptionID, credentials, &options)
+	zones, err := discoverDNSZones(context.Background(), config, credentials)
 	if err != nil {
-		return nil, err
-	}
-
-	recordClient, err := armprivatedns.NewRecordSetsClient(config.SubscriptionID, credentials, &options)
-	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("discover DNS zones: %w", err)
 	}
 
 	return &DNSProviderPrivate{
-		config:       config,
-		zoneClient:   zoneClient,
-		recordClient: recordClient,
+		config:                config,
+		credentials:           credentials,
+		serviceDiscoveryZones: zones,
 	}, nil
 }
 
@@ -57,18 +47,23 @@ func (d *DNSProviderPrivate) Present(domain, _, keyAuth string) error {
 	ctx := context.Background()
 	info := dns01.GetChallengeInfo(domain, keyAuth)
 
-	zone, err := d.getHostedZoneID(ctx, info.EffectiveFQDN)
+	zone, err := d.getHostedZone(info.EffectiveFQDN)
 	if err != nil {
 		return fmt.Errorf("azuredns: %w", err)
 	}
 
-	subDomain, err := dns01.ExtractSubDomain(info.EffectiveFQDN, zone)
+	client, err := newPrivateZoneClient(zone, d.credentials, d.config.Environment)
+	if err != nil {
+		return fmt.Errorf("azuredns: %w", err)
+	}
+
+	subDomain, err := dns01.ExtractSubDomain(info.EffectiveFQDN, zone.Name)
 	if err != nil {
 		return fmt.Errorf("azuredns: %w", err)
 	}
 
 	// Get existing record set
-	rset, err := d.recordClient.Get(ctx, d.config.ResourceGroup, zone, armprivatedns.RecordTypeTXT, subDomain, nil)
+	resp, err := client.Get(ctx, subDomain)
 	if err != nil {
 		var respErr *azcore.ResponseError
 		if !errors.As(err, &respErr) || respErr.StatusCode != http.StatusNotFound {
@@ -77,32 +72,22 @@ func (d *DNSProviderPrivate) Present(domain, _, keyAuth string) error {
 	}
 
 	// Construct unique TXT records using map
-	uniqRecords := map[string]struct{}{info.Value: {}}
-	if rset.RecordSet.Properties != nil && rset.RecordSet.Properties.TxtRecords != nil {
-		for _, txtRecord := range rset.RecordSet.Properties.TxtRecords {
-			// Assume Value doesn't contain multiple strings
-			if len(txtRecord.Value) > 0 {
-				uniqRecords[deref(txtRecord.Value[0])] = struct{}{}
-			}
-		}
-	}
+	uniqRecords := privateUniqueRecords(resp.RecordSet, info.Value)
 
 	var txtRecords []*armprivatedns.TxtRecord
 	for txt := range uniqRecords {
-		txtRecord := txt
-		txtRecords = append(txtRecords, &armprivatedns.TxtRecord{Value: []*string{&txtRecord}})
+		txtRecords = append(txtRecords, &armprivatedns.TxtRecord{Value: to.SliceOfPtrs(txt)})
 	}
 
-	ttlInt64 := int64(d.config.TTL)
 	rec := armprivatedns.RecordSet{
 		Name: &subDomain,
 		Properties: &armprivatedns.RecordSetProperties{
-			TTL:        &ttlInt64,
+			TTL:        to.Ptr(int64(d.config.TTL)),
 			TxtRecords: txtRecords,
 		},
 	}
 
-	_, err = d.recordClient.CreateOrUpdate(ctx, d.config.ResourceGroup, zone, armprivatedns.RecordTypeTXT, subDomain, rec, nil)
+	_, err = client.CreateOrUpdate(ctx, subDomain, rec)
 	if err != nil {
 		return fmt.Errorf("azuredns: %w", err)
 	}
@@ -115,17 +100,22 @@ func (d *DNSProviderPrivate) CleanUp(domain, _, keyAuth string) error {
 	ctx := context.Background()
 	info := dns01.GetChallengeInfo(domain, keyAuth)
 
-	zone, err := d.getHostedZoneID(ctx, info.EffectiveFQDN)
+	zone, err := d.getHostedZone(info.EffectiveFQDN)
 	if err != nil {
 		return fmt.Errorf("azuredns: %w", err)
 	}
 
-	subDomain, err := dns01.ExtractSubDomain(info.EffectiveFQDN, zone)
+	client, err := newPrivateZoneClient(zone, d.credentials, d.config.Environment)
 	if err != nil {
 		return fmt.Errorf("azuredns: %w", err)
 	}
 
-	_, err = d.recordClient.Delete(ctx, d.config.ResourceGroup, zone, armprivatedns.RecordTypeTXT, subDomain, nil)
+	subDomain, err := dns01.ExtractSubDomain(info.EffectiveFQDN, zone.Name)
+	if err != nil {
+		return fmt.Errorf("azuredns: %w", err)
+	}
+
+	_, err = client.Delete(ctx, subDomain)
 	if err != nil {
 		return fmt.Errorf("azuredns: %w", err)
 	}
@@ -134,21 +124,67 @@ func (d *DNSProviderPrivate) CleanUp(domain, _, keyAuth string) error {
 }
 
 // Checks that azure has a zone for this domain name.
-func (d *DNSProviderPrivate) getHostedZoneID(ctx context.Context, fqdn string) (string, error) {
-	if zone := env.GetOrFile(EnvZoneName); zone != "" {
-		return zone, nil
-	}
-
-	authZone, err := dns01.FindZoneByFqdn(fqdn)
+func (d *DNSProviderPrivate) getHostedZone(fqdn string) (ServiceDiscoveryZone, error) {
+	authZone, err := getAuthZone(fqdn)
 	if err != nil {
-		return "", fmt.Errorf("could not find zone: %w", err)
+		return ServiceDiscoveryZone{}, err
 	}
 
-	zone, err := d.zoneClient.Get(ctx, d.config.ResourceGroup, dns01.UnFqdn(authZone), nil)
+	azureZone, exists := d.serviceDiscoveryZones[dns01.UnFqdn(authZone)]
+	if !exists {
+		return ServiceDiscoveryZone{}, fmt.Errorf("could not find zone (from discovery): %s", authZone)
+	}
+
+	return azureZone, nil
+}
+
+// privateZoneClient provides Azure client for one DNS zone.
+type privateZoneClient struct {
+	zone         ServiceDiscoveryZone
+	recordClient *armprivatedns.RecordSetsClient
+}
+
+// newPrivateZoneClient creates privateZoneClient structure with initialized Azure client.
+func newPrivateZoneClient(zone ServiceDiscoveryZone, credential azcore.TokenCredential, environment cloud.Configuration) (*privateZoneClient, error) {
+	options := &arm.ClientOptions{
+		ClientOptions: azcore.ClientOptions{
+			Cloud: environment,
+		},
+	}
+
+	recordClient, err := armprivatedns.NewRecordSetsClient(zone.SubscriptionID, credential, options)
 	if err != nil {
-		return "", err
+		return nil, err
 	}
 
-	// zone.Name shouldn't have a trailing dot(.)
-	return dns01.UnFqdn(deref(zone.Name)), nil
+	return &privateZoneClient{
+		zone:         zone,
+		recordClient: recordClient,
+	}, nil
+}
+
+func (c privateZoneClient) Get(ctx context.Context, subDomain string) (armprivatedns.RecordSetsClientGetResponse, error) {
+	return c.recordClient.Get(ctx, c.zone.ResourceGroup, c.zone.Name, armprivatedns.RecordTypeTXT, subDomain, nil)
+}
+
+func (c privateZoneClient) CreateOrUpdate(ctx context.Context, subDomain string, rec armprivatedns.RecordSet) (armprivatedns.RecordSetsClientCreateOrUpdateResponse, error) {
+	return c.recordClient.CreateOrUpdate(ctx, c.zone.ResourceGroup, c.zone.Name, armprivatedns.RecordTypeTXT, subDomain, rec, nil)
+}
+
+func (c privateZoneClient) Delete(ctx context.Context, subDomain string) (armprivatedns.RecordSetsClientDeleteResponse, error) {
+	return c.recordClient.Delete(ctx, c.zone.ResourceGroup, c.zone.Name, armprivatedns.RecordTypeTXT, subDomain, nil)
+}
+
+func privateUniqueRecords(recordSet armprivatedns.RecordSet, value string) map[string]struct{} {
+	uniqRecords := map[string]struct{}{value: {}}
+	if recordSet.Properties != nil && recordSet.Properties.TxtRecords != nil {
+		for _, txtRecord := range recordSet.Properties.TxtRecords {
+			// Assume Value doesn't contain multiple strings
+			if len(txtRecord.Value) > 0 {
+				uniqRecords[deref(txtRecord.Value[0])] = struct{}{}
+			}
+		}
+	}
+
+	return uniqRecords
 }
